*: move exhaustPhysicalPlan for lock/sort/topn into physicalop
