+++ pkg/planner/core/operator/logicalop/logical_top_n.go
@@ -23,7 +23,6 @@ import (
 	ruleutil "github.com/pingcap/tidb/pkg/planner/core/rule/util"
 	"github.com/pingcap/tidb/pkg/planner/property"
 	"github.com/pingcap/tidb/pkg/planner/util"
-	"github.com/pingcap/tidb/pkg/planner/util/utilfuncp"
 	"github.com/pingcap/tidb/pkg/util/plancodec"
 )
 
@@ -148,11 +147,6 @@ func (lt *LogicalTopN) PreparePossibleProperties(_ *expression.Schema, _ ...[][]
 		return nil
 	}
 	return [][]*expression.Column{propCols}
-}
-
-// ExhaustPhysicalPlans implements base.LogicalPlan.<14th> interface.
-func (lt *LogicalTopN) ExhaustPhysicalPlans(prop *property.PhysicalProperty) ([]base.PhysicalPlan, bool, error) {
-	return utilfuncp.ExhaustPhysicalPlans4LogicalTopN(lt, prop)
 }
 
 // ExtractCorrelatedCols implements base.LogicalPlan.<15th> interface.
